This solution moves us near the top of silver league.
THis version uses voronoi tesselation all the way: the positions of enemies are placed on the playfield like an oil field, and we use voronoi to choose the best direction. When two directions have identical number of cells reachable, we use longest path (computed during tesselation).
What is lacking ? We take no care of dead players (which means the playground is full of zombies), and we don'tdetect if enemies die before us, so we don't even try to keep a path open to their deleted trace once they're dead.